requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
pymongo
requests
beautifulsoup4
google-generativeai
gunicorn